Removing global controller.
