+++ examples/device_listener.py
@@ -9,7 +9,7 @@ import sys
 import time
 
 # Import pyvera
-import pyvera
+from pyvera import VeraController
 
 
 # Define a callback that runs each time a device changes state
@@ -47,7 +47,8 @@ def main():
     args = parser.parse_args()
 
     # Start the controller
-    controller, _ = pyvera.init_controller(args.url)
+    controller = VeraController(args.url)
+    controller.start()
 
     try:
         # Get the requested device on the vera controller
