+++ examples/list_devices.py
@@ -8,7 +8,7 @@ import os
 import sys
 
 # Import pyvera
-import pyvera
+from pyvera import VeraController
 
 
 def main():
@@ -22,7 +22,8 @@ def main():
     args = parser.parse_args()
 
     # Start the controller
-    controller, _ = pyvera.init_controller(args.url)
+    controller = VeraController(args.url)
+    controller.start()
 
     try:
         # Get a list of all the devices on the vera controller
