+++ examples/show_lock_info.py
@@ -8,7 +8,7 @@ import os
 import sys
 
 # Import pyvera
-import pyvera
+from pyvera import VeraController, VeraLock
 
 
 def main():
@@ -22,7 +22,8 @@ def main():
     args = parser.parse_args()
 
     # Start the controller
-    controller, _ = pyvera.init_controller(args.url)
+    controller = VeraController(args.url)
+    controller.start()
 
     try:
         # Get a list of all the devices on the vera controller
@@ -30,7 +31,7 @@ def main():
 
         # Look over the list and find the lock devices
         for device in all_devices:
-            if isinstance(device, pyvera.VeraLock):
+            if isinstance(device, VeraLock):
                 print(
                     "{} {} ({})".format(
                         type(device).__name__, device.name, device.device_id
