+++ pyvera/const.py
@@ -1,0 +1,3 @@
+"""Constants"""
+
+TIMESTAMP_NONE = {"dataversion": 1, "loadtime": 0}
